[wip] move view_init and view_destroy to view-impl-common.c
