+++ include/view-impl-common.h
@@ -20,6 +20,9 @@ void view_impl_map(struct view *view);
  */
 void view_impl_apply_geometry(struct view *view, int w, int h);
 
+void view_init(struct view *view);
+void view_destroy(struct view *view);
+
 void view_impl_remove_common_listeners(struct view *view);
 
 #endif /* __LABWC_VIEW_IMPL_COMMON_H */
